Improve mobile layout
The looks of the site has been improved on mobile by:
- Changing the header to be more sensible tag-wise and wrap when
necessary
- Changing font sizes on different device sizes
- Removing the fixed attribute from the about-background (This doesn't
work well with cover on images in the middle of the page)
- Tweaking font colours and darkening images slightly for legibility

Additionally the name Epse has been added at the bottom with the names
of people who contributed to this site. I don't care much about it, so
if anyone feels this is inappropriate you may remove this.